refactor: 약관 조회 URI 변경에 따른 클래스 분리 - #48
